Extracted filter panel & modal into components
